requirements.txt
torch>=2.8.0
torchvision>=0.23.0
Pillow>=11.3.0
streamlit>=1.54.0
